# Improve copyto! performance