--- a/src/pool.jl
+++ b/src/pool.jl
@@ -163,7 +163,7 @@ end
 
 # Do not override Base.merge as for internal use we need to use the type and orderedness
 # of the first pool rather than promoting both pools
-function merge_pools(a::CategoricalPool{T, R}, b::CategoricalPool) where {T, R}
+function merge_pools(a::CategoricalPool{T}, b::CategoricalPool) where {T}
     if length(a) == 0 && length(b) == 0
         newlevs = T[]
         ordered = isordered(a)
@@ -177,7 +177,7 @@ function merge_pools(a::CategoricalPool{T, R}, b::CategoricalPool) where {T, R}
         nl, ordered = mergelevels(isordered(a), a.levels, b.levels)
         newlevs = convert(Vector{T}, nl)
     end
-    CategoricalPool{T, R}(newlevs, ordered)
+    newlevs, ordered
 end
 
 Base.issubset(a::CategoricalPool, b::CategoricalPool) = issubset(a.levels, keys(b.invindex))