The check was against the levels in destination before merging them with source levels, so it missed the cases where destination levels where an ordered  subset of source levels. This notably includes the case where destination levels are empty. Add even more tests to make sure we cover everything.

Avoid creating pool in merge_pools! when not needed. Creating a pool is expensive when there are many levels, and in `copyto!`we sometimes don't use it.

Avoid checking for duplicate levels when not needed. When levels have been computing internally we don't need to check for duplicates, which is expensive when the number of levels is large.

Fixes #308.


```julia
julia> using CategoricalArrays

julia> x = categorical(1:10^6);

# master
julia> y = similar(x, 10); xx=x[1:10]; @time copyto!(y, xx);
  1.053766 seconds (2.77 M allocations: 299.787 MiB, 18.95% gc time)

julia> y = similar(x, 10); xx=x[1:10]; @time copyto!(y, xx);
  0.472269 seconds (2.00 M allocations: 261.165 MiB)

julia> y = similar(x, 10); xx=x[1:10]; @time copyto!(y, xx);
  0.642487 seconds (2.00 M allocations: 261.165 MiB, 31.97% gc time)

# This PR
julia> y = similar(x, 10); xx=x[1:10]; @time copyto!(y, xx);
  0.681033 seconds (1.85 M allocations: 145.464 MiB, 22.58% gc time)

julia> y = similar(x, 10); xx=x[1:10]; @time copyto!(y, xx);
  0.153095 seconds (1.00 M allocations: 103.242 MiB)

julia> y = similar(x, 10); xx=x[1:10]; @time copyto!(y, xx);
  0.199941 seconds (1.00 M allocations: 103.242 MiB, 46.05% gc time)
```