@nalimilan 2020-10-27
I think tests cover everything (coveralls confirms it), so it should be safe. But these optimized methods are indeed very tricky...